# move PNode.comment to a side channel, reducing memory usage during compilation by a factor 1.25x