@Varriount 2021-08-28
Just to make sure I understand how this works, this change saves space because not all AST nodes have comments?