@timotheecour 2021-08-28
> Just to make sure I understand how this works, this change saves space because not all AST nodes have comments?

yes, as you can see above, ~ 1/260 PNode's have a non-empty comment field, which makes sense.

The problem right now is that the cost of the lookup is higher than i expected so it ends up being a space vs speed tradeoff unless someone has concrete ideas for how to improve this tradeoff